+++ ios/Classes/SwiftFlutterLivePlugin.swift
@@ -1,5 +1,6 @@
 import Flutter
 import UIKit
+import AVFoundation
 
 public class SwiftFlutterLivePlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
@@ -9,6 +10,29 @@ public class SwiftFlutterLivePlugin: NSObject, FlutterPlugin {
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    result("iOS " + UIDevice.current.systemVersion)
+    switch call.method {
+    case "getPlatformVersion":
+        result("iOS " + UIDevice.current.systemVersion)
+    case "setSpeakerphoneOn":
+        let args = call.arguments as! Dictionary<String,Bool>
+        if !args.keys.contains("enabled") {
+            break
+        }
+        let enabled = args["enabled"]!
+
+        let session = AVAudioSession.sharedInstance()
+        try? session.setCategory(AVAudioSession.Category.playAndRecord)
+        try? session.setMode(AVAudioSession.Mode.voiceChat)
+        if enabled {
+            try? session.overrideOutputAudioPort(AVAudioSession.PortOverride.speaker)
+        } else {
+            try? session.overrideOutputAudioPort(AVAudioSession.PortOverride.none)
+        }
+        try? session.setActive(true)
+        
+        result(nil)
+    default:
+        break
+    }
   }
 }